engine/services/faiss-recall/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
faiss-cpu==1.8.0.post1
numpy==1.26.4
